HDX-7967 switching the homepage controller to flask
including also some about pages and the count controller
