+++ ckanext-hdx_theme/ckanext/hdx_theme/splash_page.py
@@ -1,191 +1,191 @@
-import re
-from rdflib import Graph, Literal, BNode, RDF
-from rdflib.namespace import Namespace
-
-import ckan.lib.base as base
-import ckan.logic as logic
-import ckan.plugins.toolkit as tk
-import ckanext.hdx_theme.helpers.faq_wordpress as fw
-from ckan.common import config
-from ckan.controllers.home import HomeController
-from ckan.lib.base import c, h
-from ckan.lib.base import model
-from ckan.lib.base import request
-
-_ = tk._
-NotAuthorized = logic.NotAuthorized
-check_access = logic.check_access
-get_action = logic.get_action
-abort = base.abort
-redirect = tk.redirect_to
-
-
-def google_searchbox_data():
-    SCHEMA = Namespace('http://schema.org/')
-    ckan_url = config.get('ckan.site_url', 'https://data.humdata.org').replace('http://', 'https://') + '/'
-    search_action_url = ckan_url + 'search?q={search_term_string}'
-
-    website_node = BNode()
-    g = Graph()
-    g.bind('schema', SCHEMA)
-    g.add((website_node, RDF.type, SCHEMA.WebSite))
-    g.add((website_node, SCHEMA.url, Literal(ckan_url)))
-    search_action = BNode()
-
-    g.add((website_node, SCHEMA.potentialAction, search_action))
-    g.add((search_action, RDF.type, SCHEMA.SearchAction))
-    # g.add((search_action, SCHEMA.target, Literal(search_action_url)))
-    g.add((search_action, SCHEMA['query-input'], Literal('required name=search_term_string')))
-
-    entry_point = BNode()
-    g.add((search_action, SCHEMA.target, entry_point))
-    g.add((entry_point, RDF.type, SCHEMA.EntryPoint))
-    g.add((entry_point, SCHEMA.urlTemplate, Literal(search_action_url)))
-
-    return g.serialize(format='json-ld', auto_compact=True)
-    # data = '''
-    # {{
-    #   "@context": "https://schema.org",
-    #   "@type": "WebSite",
-    #   "url": "{ckan_url}",
-    #   "potentialAction": {{
-    #     "@type": "SearchAction",
-    #     "target": {{
-    #       "@type": "EntryPoint",
-    #       "urlTemplate": "{search_url_template}"
-    #     }},
-    #     "query-input": "required name=search_term_string"
-    #   }}
-    # }}
-    # '''.format(ckan_url=ckan_url, search_url_template=search_action_url)
-    # return data
-
-
-structured_data = google_searchbox_data()
-
-
-class SplashPageController(HomeController):
-    group_type = 'group'
-
-    def index(self):
-        group_type = None
-        context = {'model': model, 'session': model.Session,
-                   'user': c.user or c.author, 'for_view': True,
-                   'with_private': False}
-
-        q = c.q = request.params.get('q', '')
-        data_dict = {'all_fields': True, 'q': q}
-        sort_by = c.sort_by_selected = request.params.get('sort')
-        if sort_by:
-            data_dict['sort'] = sort_by
-        try:
-            self._check_access('site_read', context)
-        except NotAuthorized:
-            abort(403, _('Not authorized to see this page'))
-        if c.userobj:
-            context['user_id'] = c.userobj.id
-            context['user_is_admin'] = c.userobj.sysadmin
-
-        # c.group_package_stuff = caching.cached_get_group_package_stuff()
-        #
-        # ##Removing groups without geojson for the map
-        # c.group_map = []
-        # for gp in c.group_package_stuff:
-        #     '''
-        #         Removed check for geojson data because in the new version this information
-        #         does not come from the group_list action and for now we are not using the map.
-        #         If we'll need this we should implement some caching functionality for this too.
-        #     '''
-        #     c.group_map.append(gp)
-
-        if c.userobj is not None:
-            site_title = config.get('ckan.site_title', 'CKAN')
-            msg = None
-            url = h.url_for(controller='user', action='edit')
-            is_google_id = \
-                c.userobj.name.startswith('https://www.google.com/accounts/o8/id')
-            if not c.userobj.email and (is_google_id and not c.userobj.fullname):
-                msg = _(u'Please <a href="{link}">update your profile</a>'
-                        u' and add your email address and your full name. '
-                        u'{site} uses your email address'
-                        u' if you need to reset your password.'.format(
-                    link=url, site=site_title))
-            elif not c.userobj.email:
-                msg = _('Please <a href="%s">update your profile</a>'
-                        ' and add your email address. ') % url + \
-                      _('%s uses your email address'
-                        ' if you need to reset your password.') \
-                      % site_title
-            elif is_google_id and not c.userobj.fullname:
-                msg = _('Please <a href="%s">update your profile</a>'
-                        ' and add your full name.') % (url)
-            if msg:
-                h.flash_notice(msg, allow_html=True)
-
-        # template_data = {
-        #     'data': {
-        #         'hdx.carousel.config': logic.get_action('hdx_carousel_settings_show')({}, {})
-        #     }
-        # }
-        c.structured_data = structured_data
-        return base.render('home/index.html')
-
-    def _check_access(self, action_name, *args, **kw):
-        ''' select the correct group/org check_access '''
-        return check_access(self._replace_group_org(action_name), *args, **kw)
-
-    def _replace_group_org(self, string):
-        ''' substitute organization for group if this is an org'''
-        if self.group_type == 'organization':
-            string = re.sub('^group', 'organization', string)
-        return string
-
-    def _action(self, action_name):
-        ''' select the correct group/org action '''
-        return get_action(self._replace_group_org(action_name))
-
-    def about(self, page):
-        title = {
-            'license': _('Data Licenses'),
-            'terms': _('Terms of Service'),
-            'hdx-qa-process': _('HDX QA Process')
-        }
-        html = {
-            'license': 'licenses',
-            'terms': 'terms',
-            'hdx-qa-process': 'home/snippets/qa-process.html'
-        }
-        render = {
-            'hdx-qa-process': 'home/about2-light.html'
-        }
-
-        title_item = title.get(page)
-        if title_item is None:
-            abort(404, _("The requested about page doesn't exist"))
-
-        render_item = render.get(page, 'redirect')
-        if render_item == 'redirect':
-            html_item = html.get(page)
-            url = h.url_for('hdx_faqs.read', category=html_item)
-            result = redirect(url)
-            return result
-        else:
-            html_item = html.get(page)
-
-        extra_vars = {'title': title_item, 'html': html_item, 'page': page}
-        return base.render(render_item, extra_vars=extra_vars)
-
-    def about_hrinfo(self):
-        from ckan.lib.base import render_jinja2
-        title = {'hr_info': _('Legacy HR Info')}
-        html = {'hr_info': 'home/snippets/hdx_hr_info.html'}
-
-        title_item = title['hr_info']
-        html_item = html['hr_info']
-
-        if title_item is None:
-            message = _("The requested about page doesn't exist")
-            raise logic.ValidationError({'message': message}, error_summary=message)
-        html_item = render_jinja2(html_item, {})
-        extra_vars = {'title': title_item, 'html': html_item, 'page': 'hr_info'}
-        return base.render('home/about2.html', extra_vars=extra_vars)
+# import re
+# from rdflib import Graph, Literal, BNode, RDF
+# from rdflib.namespace import Namespace
+#
+# import ckan.lib.base as base
+# import ckan.logic as logic
+# import ckan.plugins.toolkit as tk
+# import ckanext.hdx_theme.helpers.faq_wordpress as fw
+# from ckan.common import config
+# from ckan.controllers.home import HomeController
+# from ckan.lib.base import c, h
+# from ckan.lib.base import model
+# from ckan.lib.base import request
+#
+# _ = tk._
+# NotAuthorized = logic.NotAuthorized
+# check_access = logic.check_access
+# get_action = logic.get_action
+# abort = base.abort
+# redirect = tk.redirect_to
+#
+#
+# def google_searchbox_data():
+#     SCHEMA = Namespace('http://schema.org/')
+#     ckan_url = config.get('ckan.site_url', 'https://data.humdata.org').replace('http://', 'https://') + '/'
+#     search_action_url = ckan_url + 'search?q={search_term_string}'
+#
+#     website_node = BNode()
+#     g = Graph()
+#     g.bind('schema', SCHEMA)
+#     g.add((website_node, RDF.type, SCHEMA.WebSite))
+#     g.add((website_node, SCHEMA.url, Literal(ckan_url)))
+#     search_action = BNode()
+#
+#     g.add((website_node, SCHEMA.potentialAction, search_action))
+#     g.add((search_action, RDF.type, SCHEMA.SearchAction))
+#     # g.add((search_action, SCHEMA.target, Literal(search_action_url)))
+#     g.add((search_action, SCHEMA['query-input'], Literal('required name=search_term_string')))
+#
+#     entry_point = BNode()
+#     g.add((search_action, SCHEMA.target, entry_point))
+#     g.add((entry_point, RDF.type, SCHEMA.EntryPoint))
+#     g.add((entry_point, SCHEMA.urlTemplate, Literal(search_action_url)))
+#
+#     return g.serialize(format='json-ld', auto_compact=True)
+#     # data = '''
+#     # {{
+#     #   "@context": "https://schema.org",
+#     #   "@type": "WebSite",
+#     #   "url": "{ckan_url}",
+#     #   "potentialAction": {{
+#     #     "@type": "SearchAction",
+#     #     "target": {{
+#     #       "@type": "EntryPoint",
+#     #       "urlTemplate": "{search_url_template}"
+#     #     }},
+#     #     "query-input": "required name=search_term_string"
+#     #   }}
+#     # }}
+#     # '''.format(ckan_url=ckan_url, search_url_template=search_action_url)
+#     # return data
+#
+#
+# structured_data = google_searchbox_data()
+#
+#
+# class SplashPageController(HomeController):
+#     group_type = 'group'
+#
+#     def index(self):
+#         group_type = None
+#         context = {'model': model, 'session': model.Session,
+#                    'user': c.user or c.author, 'for_view': True,
+#                    'with_private': False}
+#
+#         q = c.q = request.params.get('q', '')
+#         data_dict = {'all_fields': True, 'q': q}
+#         sort_by = c.sort_by_selected = request.params.get('sort')
+#         if sort_by:
+#             data_dict['sort'] = sort_by
+#         try:
+#             self._check_access('site_read', context)
+#         except NotAuthorized:
+#             abort(403, _('Not authorized to see this page'))
+#         if c.userobj:
+#             context['user_id'] = c.userobj.id
+#             context['user_is_admin'] = c.userobj.sysadmin
+#
+#         # c.group_package_stuff = caching.cached_get_group_package_stuff()
+#         #
+#         # ##Removing groups without geojson for the map
+#         # c.group_map = []
+#         # for gp in c.group_package_stuff:
+#         #     '''
+#         #         Removed check for geojson data because in the new version this information
+#         #         does not come from the group_list action and for now we are not using the map.
+#         #         If we'll need this we should implement some caching functionality for this too.
+#         #     '''
+#         #     c.group_map.append(gp)
+#
+#         if c.userobj is not None:
+#             site_title = config.get('ckan.site_title', 'CKAN')
+#             msg = None
+#             url = h.url_for(controller='user', action='edit')
+#             is_google_id = \
+#                 c.userobj.name.startswith('https://www.google.com/accounts/o8/id')
+#             if not c.userobj.email and (is_google_id and not c.userobj.fullname):
+#                 msg = _(u'Please <a href="{link}">update your profile</a>'
+#                         u' and add your email address and your full name. '
+#                         u'{site} uses your email address'
+#                         u' if you need to reset your password.'.format(
+#                     link=url, site=site_title))
+#             elif not c.userobj.email:
+#                 msg = _('Please <a href="%s">update your profile</a>'
+#                         ' and add your email address. ') % url + \
+#                       _('%s uses your email address'
+#                         ' if you need to reset your password.') \
+#                       % site_title
+#             elif is_google_id and not c.userobj.fullname:
+#                 msg = _('Please <a href="%s">update your profile</a>'
+#                         ' and add your full name.') % (url)
+#             if msg:
+#                 h.flash_notice(msg, allow_html=True)
+#
+#         # template_data = {
+#         #     'data': {
+#         #         'hdx.carousel.config': logic.get_action('hdx_carousel_settings_show')({}, {})
+#         #     }
+#         # }
+#         c.structured_data = structured_data
+#         return base.render('home/index.html')
+#
+#     def _check_access(self, action_name, *args, **kw):
+#         ''' select the correct group/org check_access '''
+#         return check_access(self._replace_group_org(action_name), *args, **kw)
+#
+#     def _replace_group_org(self, string):
+#         ''' substitute organization for group if this is an org'''
+#         if self.group_type == 'organization':
+#             string = re.sub('^group', 'organization', string)
+#         return string
+#
+#     def _action(self, action_name):
+#         ''' select the correct group/org action '''
+#         return get_action(self._replace_group_org(action_name))
+#
+#     def about(self, page):
+#         title = {
+#             'license': _('Data Licenses'),
+#             'terms': _('Terms of Service'),
+#             'hdx-qa-process': _('HDX QA Process')
+#         }
+#         html = {
+#             'license': 'licenses',
+#             'terms': 'terms',
+#             'hdx-qa-process': 'home/snippets/qa-process.html'
+#         }
+#         render = {
+#             'hdx-qa-process': 'home/about2-light.html'
+#         }
+#
+#         title_item = title.get(page)
+#         if title_item is None:
+#             abort(404, _("The requested about page doesn't exist"))
+#
+#         render_item = render.get(page, 'redirect')
+#         if render_item == 'redirect':
+#             html_item = html.get(page)
+#             url = h.url_for('hdx_faqs.read', category=html_item)
+#             result = redirect(url)
+#             return result
+#         else:
+#             html_item = html.get(page)
+#
+#         extra_vars = {'title': title_item, 'html': html_item, 'page': page}
+#         return base.render(render_item, extra_vars=extra_vars)
+#
+#     def about_hrinfo(self):
+#         from ckan.lib.base import render_jinja2
+#         title = {'hr_info': _('Legacy HR Info')}
+#         html = {'hr_info': 'home/snippets/hdx_hr_info.html'}
+#
+#         title_item = title['hr_info']
+#         html_item = html['hr_info']
+#
+#         if title_item is None:
+#             message = _("The requested about page doesn't exist")
+#             raise logic.ValidationError({'message': message}, error_summary=message)
+#         html_item = render_jinja2(html_item, {})
+#         extra_vars = {'title': title_item, 'html': html_item, 'page': 'hr_info'}
+#         return base.render('home/about2.html', extra_vars=extra_vars)
