+++ ckanext-hdx_theme/ckanext/hdx_theme/templates/error_document_template.html
@@ -40,7 +40,7 @@
           </p>
         </div>
         <div>
-          <a href="{{ h.url('hdx_home') }}" class="btn btn-primary hdx-btn btn-large">Back to homepage</a>
+          <a href="{{ h.url('hdx_splash.index') }}" class="btn btn-primary hdx-btn btn-large">Back to homepage</a>
         </div>
       </div>
