+++ ckanext-hdx_theme/ckanext/hdx_theme/templates/snippets/home_breadcrumb_item.html
@@ -1,6 +1,6 @@
 {# Used to insert the home icon into a breadcrumb #}
 <li class="home" property="itemListElement" typeof="ListItem">
-  <a href="{{ h.url('hdx_home') }}" property="item" typeof="WebPage">
+  <a href="{{ h.url('hdx_splash.index') }}" property="item" typeof="WebPage">
     <span property="name">{{ _('Home') }}</span>
   </a>
   <meta property="position" content="1">
