+++ ckanext-hdx_theme/ckanext/hdx_theme/tests/test_pages/test_page_load.py
@@ -32,8 +32,8 @@ pages = [
     # {'controller': 'ckanext.hdx_package.controllers.dataset_controller:DatasetController',
     #  'action': 'preselect', 'usertype': 'all'},
     {'controller': 'ckanext.hdx_users.controllers.dashboard_controller:DashboardController', 'action': 'read', 'has_id': True, 'usertype': 'all'},
-    {'controller': 'ckanext.hdx_theme.splash_page:SplashPageController', 'action': 'about_hrinfo', 'usertype': 'all'},
-    {'controller': 'ckanext.hdx_theme.splash_page:SplashPageController', 'action': 'index', 'usertype': 'all'}
+    {'url_name': 'hdx_splash.about_hrinfo', 'usertype': 'all'},
+    {'url_name': 'hdx_splash.index', 'usertype': 'all'}
 ]
